requirements.txt
pydantic
SpeechRecognition
thefuzz
pyaudio
setuptools
ahk>=1.7.6
json-five
pyyaml
